add crud for SupportMeasure free

+++ be/src/main/java/com/smokingcessation/repository/SupportMeasureRepository.java
@@ -3,5 +3,5 @@ package com.smokingcessation.repository;
 import com.smokingcessation.model.SupportMeasure;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupportMeasureRepository extends JpaRepository<SupportMeasure, Long> {
+public interface SupportMeasureRepository extends JpaRepository<SupportMeasure, Integer> {
 }
